Version stable 28 mars 2013

+++ app/Controller/UsersController.php
@@ -23,12 +23,12 @@
 		public function index() {
 			// $users = $this->User->find('all');
 			// $this->set('users', $users);
-			
-			if ($this->User->role == 'administrateur') {
+
+			if (AuthComponent::user('role') == 'administrateur') {
 				$this->User->recursive = 0;
 				$this->set('users', $this->paginate());
 			} elseif (AuthComponent::user('id') != null) {
-				$users = $this->User->find('first', array('conditions', array('id' => AuthComponent::user('id'))));
+				$users = $this->User->find('first', array('conditions' => array('id' => AuthComponent::user('id'))));
 				$this->set('users', $users);
 			}
 		}
@@ -43,21 +43,62 @@
 
 		public function add() {
 			if ($this->request->is('post')) {
-				$this->User->create();
-				if ($this->User->save($this->request->data)) {
-					$this->Session->setFlash(__("Le compte a été créé."));
-					$this->redirect(array('action' => 'index'));
+				$this->set('username', $_POST['data']['User']['username']);
+				$this->set('datenaissance', $_POST['data']['User']['datenaissance']);
+				$this->set('taille', $_POST['data']['User']['taille']);
+				$this->set('poids', $_POST['data']['User']['poids']);
+				$this->set('email', $_POST['data']['User']['email']);
+				$this->set('sexe', $_POST['data']['User']['sexe']);
+				$this->set('enceinte', $_POST['data']['User']['enceinte']);
+				$this->set('allaitante', $_POST['data']['User']['allaitante']);
+
+				$pseudoExiste = $this->User->find('first', array('conditions' => array('username' => $this->request->data['User']['username'])));
+
+				if (empty($pseudoExiste)) {
+					$this->request->data['User']['role'] = 'utilisateur';
+					$this->User->create();
+					if ($this->User->save($this->request->data)) {
+						$this->Session->setFlash(__("Le compte a été créé."));
+						$this->redirect(array('action' => 'index'));
+					} else {
+						$this->Session->setFlash(__("Erreur lors de la creation du compte. Merci de réessayer."));
+					}
 				} else {
-					$this->Session->setFlash(__("Erreur lors de la creation du compte. Merci de réessayer."));
+					$this->Session->setFlash(__("Le nom d'utilisateur existe déjà, veuillez en choisir un autre"));
 				}
+			} else {
+				$this->set('username', null);
+				$this->set('datenaissance', null);
+				$this->set('taille', null);
+				$this->set('poids', null);
+				$this->set('email', null);
+				$this->set('sexe', null);
+				$this->set('enceinte', null);
+				$this->set('allaitante', null);
 			}
 		}
 
 		public function edit($id = null) {
+
+			if (AuthComponent::user('role') != 'administrateur') {
+				$id = AuthComponent::user('id');
+			}
+
 			$this->User->id = $id;
+
 			if (!$this->User->exists()) {
 				throw new NotFoundException(__('Utilisateur invalide'));
 			}
+
+			$user = $this->User->find('first', array('conditions' => array('id' => $id)));
+			$this->set('username', $user['User']['username']);
+			$this->set('datenaissance', $user['User']['datenaissance']);
+			$this->set('taille', $user['User']['taille']);
+			$this->set('poids', $user['User']['poids']);
+			$this->set('email', $user['User']['email']);
+			$this->set('sexe', $user['User']['sexe']);
+			$this->set('enceinte', $user['User']['enceinte']);
+			$this->set('allaitante', $user['User']['allaitante']);
 
 			if ($this->request->is('post') || $this->request->is('put')) {
 				if ($this->User->save($this->request->data)) {
@@ -73,19 +114,33 @@
 		}
 
 		public function delete($id = null) {
-			// if (!$this->request->is('post')) {
-			// 	throw new MethodNotAllowedException();
-			// }
-			$this->User->id = $id;
-			if (!$this->User->exists()) {
-				throw new NotFoundException(__('Utilisateur invalide'));
+			
+			$this->Session->setFlash('Veuillez confirmer la suppression du compte');
+			
+			if (AuthComponent::user('role') != 'administrateur') {
+					$id = AuthComponent::user('id');
 			}
-			if ($this->User->delete()) {
-				$this->Session->setFlash(__('Utilisateur supprimé'));
+
+			$user = $this->User->find('first', array('conditions' => array('id' => $id)));
+			$this->set('user', $user);
+
+			if ($this->request->is('post')) {
+			
+				if (AuthComponent::user('role') != 'administrateur') {
+					$id = AuthComponent::user('id');
+				}
+
+				$this->User->id = $id;
+				if (!$this->User->exists()) {
+					throw new NotFoundException(__('Utilisateur invalide'));
+				}
+				if ($this->User->delete()) {
+					$this->Session->setFlash(__('Utilisateur supprimé'));
+					$this->redirect(array('action' => 'index'));
+				}
+				$this->Session->setFlash(__("L'utilisateur n'a pas pu être supprimé. Merci de réessayer."));
 				$this->redirect(array('action' => 'index'));
 			}
-			$this->Session->setFlash(__("L'utilisateur n'a pas pu être supprimé. Merci de réessayer."));
-			$this->redirect(array('action' => 'index'));
 		}
 	}
 ?>
